Gen2TeslaCharger.ino edit to add check for module-enabled status to module fault checking
Present module fault checking will reset the charger if a module has a fault even if the module is not in use. Chargers with one or two defective modules always fail but are usable. This request adds a check to see if a module with a fault is enabled or not. If not enabled, the fault is ignored.